+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ```
+2020-06-10 22:44:40 FIX: `expr_postfix()` use `loadlval()` to merge registers.
+                    Old code did not always work well with arrays.
 2020-06-09 20:06:15 FIX: Simplify cascading in `expr_assign()`.
                     Also, continue with lvalue, not rvalue.
 2020-06-09 20:44:49 FIX: Unroll `expr_land()` and `expr_lor()`.